HHH-18639 Remove support for DB2 versions older than 11.1 and start testing with DB2 12.1
