+++ nightly.Jenkinsfile
@@ -30,7 +30,7 @@ stage('Configure') {
 		new BuildEnvironment( dbName: 'mariadb_10_6' ),
 		new BuildEnvironment( dbName: 'postgresql_13' ),
 		new BuildEnvironment( dbName: 'edb_13' ),
-		new BuildEnvironment( dbName: 'db2_10_5', longRunning: true ),
+		new BuildEnvironment( dbName: 'db2_11_5' ), // Unfortunately there is no SQL Server 11.1 image, but 11.5 should mostly have feature parity
 		new BuildEnvironment( dbName: 'mssql_2017' ), // Unfortunately there is no SQL Server 2008 image, so we have to test with 2017
 // 		new BuildEnvironment( dbName: 'sybase_16' ), // There only is a Sybase ASE 16 image, so no pint in testing that nightly
 		new BuildEnvironment( dbName: 'sybase_jconn' ),
